+++ src/lib/map/index.js
