+++ src/lib/map/marker.js
