+++ src/services/search.js
@@ -2,9 +2,7 @@ const host = 'http://localhost:8080';
 const prediction = (term) => {
   return fetch(`${host}?term=${term}`)
     .then((resp) => resp.json())
-    .then((json) => {
-      console.log(json);
-    })
+    .then((json) => json)
     .catch((err) => console.error);
 };
